jsonrpclib: Fix TransportMixin implementation.
The send_request implementation in TransportMixin erroneously sent the
request body (it is only supposed to send the "POST" line and
headers). This is not a problem with py2.7, because there the
XMLTransport class from xmlrpclib invokes single_request, which is
implemented in jsonrpclib and expects this wrong behavior. The py2.6
xmlrpc lib instead invokes send_request, emits some more headers and
then invokes sent content. As the bad send_request already sent the
content, the attempt to add more headers afterwards failed.

This patch removes the send_content call from send_request and adds it
in single_request instead.