SimpleJSONRPCServer: Handle incomplete xmlrpc.server package
The "future" package under python2.7 provides the contents of
`xmlrpclib` as the `xmlrpc.server` module. This lacks the
`SimpleXMLRPCDispatcher` and `SimpleXMLRPCRequestHandler` classes
expected by this module. Try to pick the needed names from the module
within the `try` block and fall back to the python2 code if they are
missing.